Basic Redis operator
This is modeled heavily after the existing Memcached operator.

Creating a Redis will in turn create a Deployment and a Service.  The
Deployment defaults to 1 replica and there is no attempt at high
availability.  By default this uses the rhel-9/redis-6 image and
performs no additional configuration of the service.